refactor(condo): DOMA-6234 split logic of dataMappers & chart views
