+++ apps/condo/domains/analytics/components/charts/resident/ResidentByPropertyChart.tsx
@@ -4,42 +4,13 @@ import React, { useMemo } from 'react'
 import { useIntl } from '@open-condo/next/intl'
 import { Typography } from '@open-condo/ui'
 
-import PaymentChart from '@condo/domains/analytics/components/PaymentChart'
 import { PaymentChartView } from '@condo/domains/analytics/components/PaymentChartView'
 
-import type { PaymentDataType } from '@condo/domains/analytics/components/PaymentChart'
-import type { EchartsSeries } from '@condo/domains/analytics/components/TicketChart'
+import { ResidentByPropertyDataMapper } from './dataMappers'
 
-const TOP_VALUES = 9
+import type { IResidentChartCard } from './dataMappers'
 
-const ResidentByPropertyDataMapper = (residentsTitle: string) => new PaymentChart({
-    pie: {
-        chart: (viewMode, data) => {
-            const series: Array<EchartsSeries> = [{
-                name: residentsTitle,
-                data: data.slice(0, TOP_VALUES).map(resident => ({ value: resident.count, name: resident.address })),
-                radius: '75%',
-                type: 'pie',
-                label: { show: true, formatter: (e) =>  e.percent + '%' },
-            }]
-            return {
-                legend: [],
-                tooltip: { trigger: 'item', axisPointer: { type: 'none' } },
-                axisData: {
-                    yAxis: { type: 'category', data: null, axisLabel: { show: false } },
-                    xAxis: { type: 'value', data: null, boundaryGap: [0, 0.02] },
-                },
-                series,
-            }
-        },
-    },
-})
-
-interface IResidentByPropertyChart {
-    ({ data }: { data: PaymentDataType }): React.ReactElement
-}
-
-const ResidentByPropertyChart: IResidentByPropertyChart = ({ data }) => {
+const ResidentByPropertyChart: IResidentChartCard = ({ data }) => {
     const intl = useIntl()
     const ChartTitle = intl.formatMessage({ id: 'pages.reports.residentsTitle' })
     const ResidentTitle = intl.formatMessage({ id: 'global.section.contacts' })
